fix connecting wallet in mobile
